show task status
