+++ todo/tasks/task_details_renderer.py
@@ -1,6 +1,10 @@
 import humanize
 
 
+def done_task_payload(task):
+    return 'CLOSE_TASK_{}'.format(task._id)
+
+
 def open_task_payload(task):
     return 'OPEN_TASK_{}'.format(task._id)
 
@@ -13,7 +17,47 @@ def remove_task_payload(task):
     return 'REMOVE_TASK_{}'.format(task._id)
 
 
+def start_task_payload(task):
+    return 'START_TASK_{}'.format(task._id)
+
+
+def stop_task_payload(task):
+    return 'STOP_TASK_{}'.format(task._id)
+
+
 async def render(story, user, task):
+    status = getattr(task, 'status', 'Unknown')
+    if status == 'in progress':
+        buttons = [{
+            'type': 'postback',
+            'title': 'Stop',
+            'payload': stop_task_payload(task),
+        }, {
+            'type': 'postback',
+            'title': 'Done',
+            'payload': done_task_payload(task),
+        }, ]
+    elif status == 'close':
+        buttons = [{
+            'type': 'postback',
+            'title': 'Reopen',
+            'payload': reopen_task_payload(task),
+        }]
+    else:
+        buttons = [{
+            'type': 'postback',
+            'title': 'Start',
+            'payload': start_task_payload(task),
+        }]
+        # open by default
+        status = 'open'
+
+    buttons.append({
+        'type': 'postback',
+        'title': 'Remove',
+        'payload': remove_task_payload(task),
+    })
+
     await story.send_template(
         payload={
             'template_type': 'generic',
@@ -22,21 +66,9 @@ async def render(story, user, task):
                 # TODO: maybe we could generate individual images for each task
                 # 'image_url': 'https://petersfancybrownhats.com/company_image.png',
                 'subtitle': 'Status: {}\n'
-                            'Created: {}\n'.format(getattr(task, 'status', 'Unknown'),
+                            'Created: {}\n'.format(status,
                                                    humanize.naturaltime(task.created_at)),
-                'buttons': [{
-                    'type': 'postback',
-                    'title': 'Re-Open',
-                    'payload': reopen_task_payload(task),
-                }, {
-                    'type': 'postback',
-                    'title': 'Remove',
-                    'payload': remove_task_payload(task),
-                }, {
-                    'type': 'postback',
-                    'title': 'Next Task',
-                    'payload': open_task_payload(task),
-                }, ],
+                'buttons': buttons,
             }]
         },
         user=user
